+++ java/com/aionemu/gameserver/skillengine/effect/PetOrderUseUltraSkillEffect.java
@@ -1,18 +1,18 @@
-/*
- * This file is part of the Aion-Emu project.
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+/**
+ * This file is part of Aion-Lightning <aion-lightning.org>.
+ *
+ *  Aion-Lightning is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Aion-Lightning is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details. *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Aion-Lightning.
+ *  If not, see <http://www.gnu.org/licenses/>.
  */
 package com.aionemu.gameserver.skillengine.effect;
 
@@ -29,6 +29,8 @@ import com.aionemu.gameserver.skillengine.model.Effect;
 import com.aionemu.gameserver.utils.PacketSendUtility;
 
 /**
+ * Handles the visual and functional effects triggered when a pet uses an ultra skill.<br>
+ * This class processes specific logic for {@link SummonController} actions related to pet skills.
  * @author ATracer modified by Sippolo
  */
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -38,6 +40,12 @@ public class PetOrderUseUltraSkillEffect extends EffectTemplate
 	@XmlAttribute
 	protected boolean release;
 	
+	/**
+	 * Applies a specific {@code Effect} to a target.<br>
+	 * This method checks if the target is an instance of {@link Player}.<br>
+	 * It processes the logic required for the effect to take place.
+	 * @param effect The {@code Effect} object to be applied.
+	 */
 	@Override
 	public void applyEffect(Effect effect)
 	{
@@ -48,13 +56,13 @@ public class PetOrderUseUltraSkillEffect extends EffectTemplate
 			return;
 		}
 		
-		final int effectorId = effector.getSummon().getObjectId().intValue();
+		final int effectorId = effector.getSummon().getObjectId();
 		
 		final int npcId = effector.getSummon().getNpcId();
 		final int orderSkillId = effect.getSkillId();
 		
 		final int petUseSkillId = DataManager.PET_SKILL_DATA.getPetOrderSkill(orderSkillId, npcId);
-		final int targetId = effect.getEffected().getObjectId().intValue();
+		final int targetId = effect.getEffected().getObjectId();
 		
 		// Handle automatic release if skill expects so
 		if (release)
@@ -65,9 +73,16 @@ public class PetOrderUseUltraSkillEffect extends EffectTemplate
 				effector.getSummon().getController().setReleaseAfterSkill(petUseSkillId);
 			}
 		}
+		
 		PacketSendUtility.sendPacket(effector, new SM_SUMMON_USESKILL(effectorId, petUseSkillId, 1, targetId));
 	}
 	
+	/**
+	 * Calculates the attributes for a specific {@code Effect}.<br>
+	 * This method updates the {@code effect} to include an AP boost.<br>
+	 * It also links this instance as a success effect.
+	 * @param effect The {@code Effect} object to be updated.
+	 */
 	@Override
 	public void calculate(Effect effect)
 	{
